优化提交

+++ src/main/java/com/example/travel/order/dto/CreateOrderDTO.java
@@ -43,5 +43,5 @@ public class CreateOrderDTO {
     /**
      * 游客信息
      */
-    private List<Integer> touristIds;
+    private List<Long> touristIds;
 }
